chinh sua mot so xung do sau khi hop code cuoi ngay (file FoodDAO.java va FoodDetails.java)

+++ src/model/FoodDAO.java
@@ -78,7 +78,7 @@ public class FoodDAO {
         return 0;
     }
     
-    public static int update(Food old_food, String newname, String newinfo, int newprice, byte newisdrink)
+    public static int update(Food old_food, String newname, String newinfo, float newprice, byte newisdrink)
     {
         String sql = "UPDATE tbFood SET name = ?, info = ?, price = ?, isdrink = ? WHERE food_id = ?";
